% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WJD.R
\docType{data}
\name{phrase_item_bank}
\alias{phrase_item_bank}
\title{Weimar Jazz Database (WJD), phrase version}
\format{
A function containing different versions of the WJD data frame, returned with "main", "ngram", "phrases" and "files" keys
}
\source{
\url{https://jazzomat.hfm-weimar.de/dbformat/dboverview.html}
}
\usage{
phrase_item_bank
}
\description{
A dataset containing the prices and other attributes of just over 11,000 melodic jazz solo phrases
}
\keyword{datasets}
